Resolve "Output map origin bug"

Closes #72

See merge request manchester_qbi/manchester_qbi_public/madym_cxx!60

+++ madym/run/mdm_RunTools_madym_DicomConvert.h
@@ -17,7 +17,7 @@
 struct dcmNumericInfo {
   int seriesNumber; ///< series number
   int  acquisitionNumber; ///< acquisitionNumber
-  int  temporalPositionIdentifier=1;  ///< temporalPositionIdentifier
+  double  temporalPositionIdentifier=1;  ///< temporalPositionIdentifier
   double  sliceLocation;  ///< sliceLocation
   int instanceNumber;  ///< instanceNumber
 };
